requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
distro==1.8.0
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.3.3
outcome==1.3.0.post0
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyPDF2==3.0.1
pypdfium2==4.24.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
